Remove column wraparound
Moving left or right at an edge column onto a non-existent column will
keep position on current column, instead of doing a wraparound like I
thought it should previously.

Got rid of function to move to next column that has stories, but now I
realise that I will need it back in order to finish arrow movement.
What if there's a column in the middle that's empty? You still need to
get to the next column that has stories. We'll have to work that out.